requirements.txt
boto3==1.35.34
python-dotenv==1.0.1
paramiko==4.0.0